Signup, Signin, ForgotPassword realtime db set up successful
For other pages to use:
import { auth } from "../firebase";

To get UID:
auth.currentUser.uid
